Move creation of tray icon to ClientUI
Move creation of tray icon from RuneLite to ClientUI.
